+++ tests/components/samples/HSS_fccAl_DGSSXRes_plate.py
